Move ContinuingTimeLimit to a new file
